requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
google-genai>=1.0.0
pydantic==2.7.1
PyMuPDF==1.24.9
slowapi==0.1.9
streamlit==1.32.0
requests==2.31.0
sqlalchemy==2.0.29
pytest==8.1.1
httpx==0.27.0
